replace TextMetric with more generic WrappedMetricWithUnbatchFunction and move metrics into common package
