+++ src/pie_modules/taskmodules/common/__init__.py
@@ -1,2 +1,3 @@
+from .metrics import WrappedMetricWithUnbatchFunction
 from .mixins import BatchableMixin
 from .utils import get_first_occurrence_index
