--- src/pie_modules/taskmodules/metrics/__init__.py
@@ -1,1 +1,0 @@
-from .text import TextMetric
